added test for extract-compact-dbg.py
