+++ tests/test_sandbox_scripts.py
@@ -326,3 +326,16 @@ def test_multirename_fasta():
     _, out, err = utils.runscript('multi-rename.py', args, sandbox=True)
     r = open(infile2).read()
     assert r in out
+
+
+def test_extract_compact_dbg_1():
+    infile = utils.get_test_data('simple-genome.fa')
+    outfile = utils.get_temp_filename('out.gml')
+    args = ['-x', '1e4', '-o', outfile, infile]
+    _, out, err = utils.runscript('extract-compact-dbg.py', args, sandbox=True)
+
+    print(out)
+    print(err)
+    assert os.path.exists(outfile)
+
+    assert '174 segments, containing 2803 nodes' in out
